delete: unused trait, impl import
