+++ domain/src/traits/repositorys.rs
@@ -1,4 +1,4 @@
-use crate::models::guild_data::{OtherGuild, OtherTimes, OwnGuild, OwnTimes};
+use crate::models::guild_data::{OtherTimes, OwnGuild, OwnTimes};
 
 // type Result<T>;
 // で関連型を使うのがうまくいくのかどうか調べるため，ここだけそうしている
@@ -30,29 +30,6 @@ pub trait OwnTimesRepository {
         &self,
         member_id: u64,
     ) -> impl std::future::Future<Output = Self::Result<Option<OwnTimes>>> + Send;
-}
-
-pub trait OtherGuildRepository {
-    // 正直asyncはいらない
-    type Result<T>;
-
-    fn upsert(
-        &self,
-        other_server: OtherGuild,
-    ) -> impl std::future::Future<Output = Self::Result<OtherGuild>> + Send;
-    fn get(
-        &self,
-        server_name: &str,
-    ) -> impl std::future::Future<Output = Self::Result<Option<OtherGuild>>> + Send;
-    fn get_all(&self) -> impl std::future::Future<Output = Self::Result<Vec<OtherGuild>>> + Send;
-    fn get_from_guild_id(
-        &self,
-        guild_id: u64,
-    ) -> impl std::future::Future<Output = Self::Result<Option<OtherGuild>>> + Send;
-    fn delete(
-        &self,
-        server_name: &str,
-    ) -> impl std::future::Future<Output = Self::Result<OtherGuild>> + Send;
 }
 
 pub trait OtherTimesRepository {
