+++ infrastructure/sled_repository/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod error;
-pub mod other_guild_repository;
 pub mod other_times_repository;
 pub mod own_guild_repository;
 pub mod own_times_repository;
