+++ presenter/discord/commands/src/global_data.rs
@@ -3,7 +3,6 @@ use anyhow::Error;
 use ca_driver::my_ca_driver::MyCaDriver;
 use domain::traits::communicators::{GuildName, HashKey};
 use signer_verifier::key_generator::RsaKeyGenerator;
-use sled_repository::other_guild_repository::SledOtherGuildRepository;
 use sled_repository::other_times_repository::SledOtherTimesRepository;
 use sled_repository::own_guild_repository::SledOwnGuildRepository;
 use sled_repository::own_times_repository::SledOwnTimesRepository;
@@ -31,7 +30,6 @@ pub struct Data {
     // データ保存にかかわるオブジェクト
     pub own_server_repository: Arc<SledOwnGuildRepository>,
     pub own_times_repository: Arc<SledOwnTimesRepository>,
-    pub other_server_repository: Arc<SledOtherGuildRepository>,
     pub other_times_repository: Arc<SledOtherTimesRepository>,
     // 鍵に関わるオブジェクト
     pub ubiquitimes_keygenerator: Arc<RsaKeyGenerator>,
